% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsAccounts.R
\name{analytics.account.profiles}
\alias{analytics.account.profiles}
\title{analytics.account.profiles}
\usage{
analytics.account.profiles(account, context = FALSE)
}
\arguments{
\item{account}{String OAuth token}

\item{context}{Boolean provide extra context}
}
\description{
List profiles available to account
}

