% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsConnect.R
\name{analytics.connect}
\alias{analytics.connect}
\title{analytics.connect}
\usage{
analytics.connect(json = NULL, jsonText = NULL)
}
\arguments{
\item{json}{File location}

\item{jsonText}{String representation of JSON}
}
\description{
Connect to Google API OAuth
}

