% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsListDimensions.R
\name{analytics.list.dimensions}
\alias{analytics.list.dimensions}
\title{Analytics account dimenstions}
\usage{
analytics.list.dimensions(accessToken, account = NULL, accountID = NULL,
  webID = NULL, startIndex = 1, maxResults = 10)
}
\arguments{
\item{accessToken}{OAUTH Token}

\item{account}{Full account to pull ID and webID from}

\item{accountID}{account ID string}

\item{webID}{web ID string}

\item{startIndex}{Integer}

\item{maxResults}{Integer}
}
\description{
Find all of the custom dimensions associated with the 
account.
}

