% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsQuery.R
\name{analytics.query}
\alias{analytics.query}
\title{analytics.query}
\usage{
analytics.query(accessToken = NULL, ids = NULL, startDate = "30daysAgo",
  endDate = "today", dimensions = c("ga:eventAction"),
  metrics = c("ga:eventValue"), columnNames = NULL, uniqueBy = NULL,
  startIndex = 1, maxResults = 10000, appendEvents = TRUE)
}
\arguments{
\item{accessToken}{OAUTH Token}

\item{ids}{String}

\item{startDate}{starting date for query results}

\item{endDate}{end date for query results}

\item{dimensions}{List of dimensions to query}

\item{metrics}{List of metrics to query}

\item{columnNames}{List of names to put on the results}

\item{uniqueBy}{List holding columns that indicate uniqueness}

\item{startIndex}{Integer index of which to start}

\item{maxResults}{Integer for total number of results}

\item{appendEvents}{Boolean}
}
\description{
Query the Analytics API showing the first 10,000 results from
the last 30 days.
}

