% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analytics.R
\name{GA}
\alias{GA}
\title{Main function for connection to Google Analytics}
\usage{
GA(json = NULL, jsonText = NULL, set_global = FALSE,
  set_global_var = "GA")
}
\description{
Main function for connection to Google Analytics
}

