% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DateUtilities.R
\name{MilliSecondsToDate}
\alias{MilliSecondsToDate}
\title{Convert JS Milliseconds to DateTime}
\usage{
MilliSecondsToDate(x, timezone = "GMT")
}
\description{
Convert JS Milliseconds to DateTime
}

