% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analytics.R
\name{GA}
\alias{GA}
\title{Main function for connection to Google Analytics}
\usage{
GA(json = NULL, jsonText = NULL, set_global = FALSE,
  set_global_var = "GA")
}
\arguments{
\item{json}{String file location to load JSON data from}

\item{jsonText}{String json representation}

\item{set_global}{Boolean}

\item{set_global_var}{String to use for name of the global}
}
\description{
Wrapper for Analytics functions
}

